src/Attacker/malicious/Backend/requirements.txt
# Backend Server Dependencies
flask>=2.3.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
